order도메인에서 호출하는 api 생성

+++ src/main/java/com/sesac/foodtruckitem/infrastructure/persistence/mysql/entity/Item.java
@@ -20,7 +20,7 @@ public class Item extends BaseEntity {
     private Long id;                // 
     private String name;            // 메뉴명 
     private String description;     // 설명
-    private int price;              // 가격
+    private long price;             // 가격
 
     @Embedded
     private Images itemImg;
